backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
sqlalchemy
pydantic-settings
pandas
numpy
python-docx
requests
python-magic
pdfplumber
openpyxl
pip install python-dotenv



fastapi>=0.68.0
uvicorn>=0.15.0
pandas>=1.3.0
numpy>=1.21.0
python-docx>=0.8.10
requests>=2.26.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0